Refactor analyze processing boundaries
Split the analyze command so command.py keeps CLI and presentation,
process.py owns Reaper project parsing and generic plugin handling, and
Output Arcade support lives in a dedicated vendor-specific module.

This makes the Reaper-specific responsibilities more explicit without
changing CLI behavior, and keeps the existing analyze test scenarios
working with only minimal import-path updates.
