+++ src/music/commands/analyze/output/__init__.py
@@ -1,0 +1,1 @@
+"""Vendor-specific analyze helpers."""
